# docs/swagger.yaml
basePath: /api
definitions:
  RIP_internal_app_ds.CartBadgeDTO:
    properties:
      count:
        type: integer
      request_id:
        type: integer
    type: object
...
  RIP_internal_app_ds.ComponentInRequest:
    properties:
      component_id:
        type: integer
      count:
        type: integer
      description:
        type: string
      image_url:
        type: string
      specifications:
        items:
          type: string
        type: array
      tdp:
        type: integer
      title:
        type: string
    type: object
  RIP_internal_app_ds.ComponentToRequestUpdateRequest:
    properties:
      count:
        type: integer
    type: object
...
  RIP_internal_app_ds.CoolRequestDTO:
    properties:
      completion_date:
        type: string
      components:
        items:
          $ref: '#/definitions/RIP_internal_app_ds.ComponentInRequest'
        type: array
      cooling_power:
        type: number
      creation_date:
        type: string
      creator_id:
        type: integer
      forming_date:
        type: string
      id:
        type: integer
      moderator_id:
        type: integer
      room_area:
        type: number
      room_height:
        type: number
      status:
        type: integer
    type: object
  RIP_internal_app_ds.CoolRequestResolveRequest:
    properties:
      action:
        description: '"complete" | "reject"'
        type: string
    required:
    - action
    type: object
  RIP_internal_app_ds.CoolRequestUpdateRequest:
    properties:
      room_area:
        type: number
      room_height:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления заявками и компонентами серверов в системе
    CoolRequest.
  title: API для системы FRAX
  version: "1.0"
paths:
...
  /coolrequests:
    get:
      description: Возвращает отфильтрованный список всех сформированных заявок (кроме
        черновиков и удаленных).
      parameters:
      - description: Фильтр по статусу заявки
        in: query
        name: status
        type: integer
      - description: Фильтр по дате 'от' (формат YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Фильтр по дате 'до' (формат YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/RIP_internal_app_ds.CoolRequestDTO'
            type: array
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список заявок (авторизованный пользователь)
      tags:
      - coolrequests
  /coolrequests/{id}:
    delete:
      description: Логически удаляет заявку, переводя ее в статус "удалена".
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить заявку (авторизованный пользователь)
      tags:
      - coolrequests
    get:
      description: Возвращает полную информацию о заявке, включая привязанные компоненты.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RIP_internal_app_ds.CoolRequestDTO'
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить одну заявку по ID (авторизованный пользователь)
      tags:
      - coolrequests
    put:
      consumes:
      - application/json
      description: Позволяет пользователю обновить поля своей заявки (площадь помещения,
        высота помещения).
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/RIP_internal_app_ds.CoolRequestUpdateRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить данные заявки (авторизованный пользователь)
      tags:
      - coolrequests
  /coolrequests/{id}/components/{component_id}:
    delete:
      description: Удаляет связь между заявкой и компонентом.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID компонента
        in: path
        name: component_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить компонент из заявки (авторизованный пользователь)
      tags:
      - m-m
    put:
      consumes:
      - application/json
      description: Изменяет количество конкретного компонента в рамках одной заявки.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID компонента
        in: path
        name: component_id
        required: true
        type: integer
      - description: Новое количество
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/RIP_internal_app_ds.ComponentToRequestUpdateRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить количество компонента в заявке (авторизованный пользователь)
      tags:
      - m-m
  /coolrequests/{id}/form:
    put:
      description: Переводит заявку из статуса "черновик" в "сформирована" и рассчитывает
        мощность охлаждения.
      parameters:
      - description: ID заявки (черновика)
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Не все поля заполнены
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать заявку (авторизованный пользователь)
      tags:
      - coolrequests
  /coolrequests/{id}/resolve:
    put:
      consumes:
      - application/json
      description: Модератор завершает или отклоняет заявку системы охлаждения.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: 'Действие: ''complete'' или ''reject'''
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/RIP_internal_app_ds.CoolRequestResolveRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить или отклонить заявку (только модератор)
      tags:
      - coolrequests
  /coolrequests/coolcart:
    get:
      description: Возвращает ID черновика текущего пользователя и количество компонентов
        в нем.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RIP_internal_app_ds.CartBadgeDTO'
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить информацию для иконки корзины (авторизованный пользователь)
      tags:
      - coolrequests
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"